refactor(core): extract shared test helpers for harness tests
